fix(core): SQLite REAL is an 8-byte float, not a 4-byte one
`sql_type_to_neutral` and `datatype_to_neutral` applied PostgreSQL's
widths to every engine, mapping `REAL`/`Float4` to `float32`. That is
right for PostgreSQL, where `real` is `float4`, and wrong for SQLite,
whose REAL storage class is defined as an 8-byte IEEE float — SQLite has
no 4-byte float at all.

This release made it urgent rather than merely latent: the aggregate
inference fix moved `SELECT SUM(total)` over a `REAL NOT NULL` column off
a decimal type, so the regenerated output had started emitting `f32` in
Rust and `float`/`getFloat` in Java where it previously had wider types —
a newly introduced silent narrowing.

`Catalog` now carries the dialect it was built from, so both conversion
functions became dialect-aware without changing any signature: every
caller already passes `&Catalog`. `sql_type` strings stay dialect-neutral
and the width decision happens where the other type semantics are already
resolved.

Only the four statically typed SQLite targets move, all float to double.
PostgreSQL's `real` is unchanged and asserted so in the same test.

SQLite's `INTEGER` has the same problem — it holds up to 8 bytes and is
mapped to `int32` — but that would retype every SQLite primary key and is
not newly introduced, so it stays on #70.

+++ integration_tests/rust-sqlx-sqlite/src/queries.rs
@@ -4,14 +4,14 @@
 #[derive(Debug, Clone, sqlx::FromRow)]
 pub struct GetOrdersByUserRow {
     pub id: i32,
-    pub total: f32,
+    pub total: f64,
     pub notes: Option<String>,
     pub created_at: String,
 }
 
 #[derive(Debug, Clone, sqlx::FromRow)]
 pub struct GetOrderTotalRow {
-    pub total_sum: Option<f32>,
+    pub total_sum: Option<f64>,
 }
 
 #[derive(Debug, Clone, sqlx::FromRow)]
